requirements.txt
asgiref==3.5.2
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.13
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.7.7
pycodestyle==2.8.0
PyJWT==2.4.0
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.2.0
whitenoise==6.2.0
